Set English as the default to allow automated translation services
This pr converts the following languages into English. The idea is to set the repo as a default of English and allow translations of other languages. We are using an automated translation service where we need the base line default files to be constant and set as a single language. As 99.99% of the repo is set to English we have decided to make it 100% English.

The following translations took place in this pr:

- French to English
- Japanese to English
- Chinese to English
- Russian to English
- North Korean to English

Also there are a few coding mistakes cleaned up such as `”` to this `"`

This pr also merges the English and Chinese Surf Browser pr into this one plus convert the Chinese name into English. See here old pr (not yet merged and moved into this one): https://github.com/WhichBrowser/Parser-PHP/pull/332

Note:

Can't test "Open-Sankore" user agent as I can't find one online and their website has closed down.